reorganize structure w/ functional edit button
